+++ docs/changelog.md
@@ -1,3 +1,33 @@
+## [3.0.0] - June 26 2021
+### Added
+- ``Zone:trackItem(instance)``
+- ``Zone:untrackItem(instance)``
+- ``Zone.itemEntered`` event
+- ``Zone.itemExited`` event
+- ``ZoneController.setGroup(settingsGroupName, properties)``
+- ``ZoneController.getGroup(settingsGroupName)``
+- ``SettingsGroup.onlyEnterOnceExitedAll`` property
+- ``Zone:bindToGroup(settingsGroupName)``
+- ``Zone:unbindFromGroup(settingsGroupName)``
+- ``Zone:findPoint(position)``
+- ``ZoneController.getCharacterSize(player)``
+
+### Changed
+- Internal behaviour to use the new Spatial Query API (https://devforum.roblox.com/t/introducing-overlapparams-new-spatial-query-api/1435720) instead of the Region3 API.
+- The default Detection from ``Automatic`` to ``Centre``.
+- The behaviour of Detection ``Centre`` to include the whole HumanoidRootPart instead of a singular Vector within (this was required due to the new Spatial Query API).  
+- ``Zone:findPart`` now returns array ``touchingZoneParts`` as its second value.
+- ``Maid`` to [``Janitor``](https://github.com/howmanysmall/Janitor) by howmanysmall.
+
+### Removed
+- RotatedRegion3
+- ``ZoneController.getCharacterRegion``
+- ``ZoneController.verifyTouchingParts``
+- ``ZoneController.vectorIsBetweenYBounds``
+
+
+
+--------
 ## [2.2.3] - June 17 2021
 ### Fixed
 - The incorrect disabling of Seats and VehicleSeats within Part Zones.